Added Changelogs, Examples, and reformatting API (#17647)
* Changelog + Reformatting

* Disable Tests
