+++ sdk/storage/azblob/doc.go
@@ -28,7 +28,7 @@ To create a client object, you will need the account's blob service endpoint URL
 Types of Credentials
 
 The clients support different forms of authentication.
-The azblob library supports any of the `azcore.TokenCredential` interfaces, authorization via a connection String,
+The azblob library supports any of the `azcore.TokenCredential` interfaces, authorization via a Connection String,
 or authorization with a Shared Access Signature token.
 
 Using a Shared Key
@@ -46,13 +46,18 @@ Use the key as the credential parameter to authenticate the client:
 	if !ok {
 		panic("AZURE_STORAGE_ACCOUNT_KEY could not be found")
 	}
-	credential, err := NewSharedKeyCredential(accountName, accountKey)
+
+    serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+
+	cred, err := azblob.NewSharedKeyCredential(accountName, accountKey)
 	handle(err)
 
-	serviceClient, err := azblob.NewServiceClient("https://<my_account_name>.blob.core.windows.net/", cred, nil)
+    serviceClient, err := azblob.NewServiceClientWithSharedKey(serviceURL, cred, nil)
 	handle(err)
 
-Using a connection String
+    fmt.Println(serviceClient.URL())
+
+Using a Connection String
 
 Depending on your use case and authorization method, you may prefer to initialize a client instance with a connection string instead of providing the account URL and credential separately.
 To do this, pass the connection string to the service client's `NewServiceClientFromConnectionString` method.
@@ -64,7 +69,7 @@ The connection string can be found in your storage account in the Azure Portal u
 Using a Shared Access Signature (SAS) Token
 
 To use a shared access signature (SAS) token, provide the token at the end of your service URL.
-You can generate a SAS token from the Azure Portal under Shared Access Signature or use the ServiceClient.GetSASURL() functions.
+You can generate a SAS token from the Azure Portal under Shared Access Signature or use the ServiceClient.GetSASToken() functions.
 
 	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
 	if !ok {
@@ -74,28 +79,27 @@ You can generate a SAS token from the Azure Portal under Shared Access Signature
 	if !ok {
 		panic("AZURE_STORAGE_ACCOUNT_KEY could not be found")
 	}
-	credential, err := azblob.NewSharedKeyCredential(accountName, accountKey)
+	serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+
+	cred, err := azblob.NewSharedKeyCredential(accountName, accountKey)
+	handle(err)
+	serviceClient, err := azblob.NewServiceClientWithSharedKey(serviceURL, cred, nil)
+	handle(err)
+	fmt.Println(serviceClient.URL())
+
+	// Alternatively, you can create SAS on the fly
+
+	resources := azblob.AccountSASResourceTypes{Service: true}
+	permission := azblob.AccountSASPermissions{Read: true}
+	start := time.Now()
+	expiry := start.AddDate(0, 0, 1)
+	serviceURLWithSAS, err := serviceClient.GetSASURL(resources, permission, start, expiry)
 	handle(err)
 
-	serviceClient, err := azblob.NewServiceClient(fmt.Sprintf("https://%s.blob.core.windows.net/", accountName), credential, nil)
+	serviceClientWithSAS, err := azblob.NewServiceClientWithNoCredential(serviceURLWithSAS, nil)
 	handle(err)
 
-    // Provide the convenience function with relevant info
-	accountSAS, err := serviceClient.GetSASURL(AccountSASResourceTypes{Object: true, Service: true, Container: true}, AccountSASPermissions{Read: true, List: true}, AccountSASServices{Blob: true}, time.Now(), time.Now().Add(48*time.Hour))
-	handle(err)
-
-	urlToSend := fmt.Sprintf("https://%s.blob.core.windows.net/?%s", accountName, accountSAS)
-	// You can hand off this URL to someone else via any mechanism you choose.
-
-	// ******************************************
-
-	// When someone receives the URL, they can access the resource using it in code like this, or a tool of some variety.
-	serviceClient, err = azblob.NewServiceClient(urlToSend, azcore.NewAnonymousCredential(), nil)
-	handle(err)
-
-	// You can also break a blob URL up into its constituent parts
-	blobURLParts := azblob.NewBlobURLParts(serviceClient.URL())
-	fmt.Printf("SAS expiry time = %s\n", blobURLParts.SAS.ExpiryTime())
+	fmt.Println(serviceClientWithSAS.URL())
 
 Types of Clients
 
@@ -105,89 +109,91 @@ There are three different clients provided to interact with the various componen
     * Get and set account settings.
     * Query, create, and delete containers within the account.
 
-2. **`containerClient`**
+2. **`ContainerClient`**
     * Get and set container access settings, properties, and metadata.
     * Create, delete, and query blobs within the container.
     * `ContainerLeaseClient` to support container lease management.
 
 3. **`BlobClient`**
-    * `appendBlobClient`, `BlockBlobClient`, and `PageBlobClient`
+    * `AppendBlobClient`, `BlockBlobClient`, and `PageBlobClient`
     * Get and set blob properties.
     * Perform CRUD operations on a given blob.
     * `BlobLeaseClient` to support blob lease management.
 
 Examples
 
-	// Use your storage account's name and key to create a credential object, used to access your account.
-	// You can obtain these details from the Azure Portal.
+	// Your account name and key can be obtained from the Azure Portal.
 	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
 	if !ok {
-		handle(errors.New("AZURE_STORAGE_ACCOUNT_NAME could not be found"))
+		panic("AZURE_STORAGE_ACCOUNT_NAME could not be found")
 	}
 
 	accountKey, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_KEY")
 	if !ok {
-		handle(errors.New("AZURE_STORAGE_ACCOUNT_KEY could not be found"))
+		panic("AZURE_STORAGE_ACCOUNT_KEY could not be found")
 	}
-	cred, err := NewSharedKeyCredential(accountName, accountKey)
+	cred, err := azblob.NewSharedKeyCredential(accountName, accountKey)
 	handle(err)
 
-	// Open up a service client.
-	// You'll need to specify a service URL, which for blob endpoints usually makes up the syntax http(s)://<account>.blob.core.windows.net/
-	service, err := NewServiceClient(fmt.Sprintf("https://%s.blob.core.windows.net/", accountName), cred, nil)
+	// The service URL for blob endpoints is usually in the form: http(s)://<account>.blob.core.windows.net/
+	serviceClient, err := azblob.NewServiceClientWithSharedKey(fmt.Sprintf("https://%s.blob.core.windows.net/", accountName), cred, nil)
 	handle(err)
 
-	// All operations in the Azure Blob Storage SDK for Go operate on a context.Context, allowing you to control cancellation/timeout.
-	ctx := context.Background() // This example has no expiry.
+	// ===== 1. Create a container =====
 
-	// This example showcases several common operations to help you get started, such as:
-
-	// ===== 1. Creating a container =====
-
-	// First, branch off of the service client and create a container client.
-	container := service.NewContainerClient("mycontainer")
-	// Then, fire off a create operation on the container client.
-	// Note that, all service-side requests have an options bag attached, allowing you to specify things like metadata, public access types, etc.
-	// Specifying nil omits all options.
-	_, err = container.Create(ctx, nil)
+	// First, create a container client, and use the Create method to create a new container in your account
+	containerClient, err := serviceClient.NewContainerClient("testcontainer")
 	handle(err)
 
-	// ===== 2. Uploading/downloading a block blob =====
-	// We'll specify our data up-front, rather than reading a file for simplicity's sake.
-	data := "Hello world!"
+	// All APIs have an options' bag struct as a parameter.
+	// The options' bag struct allows you to specify optional parameters such as metadata, public access types, etc.
+	// If you want to use the default options, pass in nil.
+	_, err = containerClient.Create(context.TODO(), nil)
+	handle(err)
 
-	// Branch off of the container into a block blob client
-	blockBlob := container.NewBlockBlobClient("HelloWorld.txt")
+	// ===== 2. Upload and Download a block blob =====
+	uploadData := "Hello world!"
+
+	// Create a new blockBlobClient from the containerClient
+	blockBlobClient, err := containerClient.NewBlockBlobClient("HelloWorld.txt")
+	handle(err)
 
 	// Upload data to the block blob
-	_, err = blockBlob.Upload(ctx, NopCloser(strings.NewReader(data)), nil)
+	blockBlobUploadOptions := azblob.BlockBlobUploadOptions{
+		Metadata: map[string]string{"Foo": "Bar"},
+		TagsMap:  map[string]string{"Year": "2022"},
+	}
+	_, err = blockBlobClient.Upload(context.TODO(), streaming.NopCloser(strings.NewReader(uploadData)), &blockBlobUploadOptions)
 	handle(err)
 
 	// Download the blob's contents and ensure that the download worked properly
-	get, err := blockBlob.Download(ctx, nil)
+	blobDownloadResponse, err := blockBlobClient.Download(context.TODO(), nil)
 	handle(err)
 
-	// Open a buffer, reader, and then download!
-	downloadedData := &bytes.Buffer{}
-	reader := get.Body(RetryReaderOptions{}) // RetryReaderOptions has a lot of in-depth tuning abilities, but for the sake of simplicity, we'll omit those here.
-	_, err = downloadedData.ReadFrom(reader)
+	// Use the bytes.Buffer object to read the downloaded data.
+	// RetryReaderOptions has a lot of in-depth tuning abilities, but for the sake of simplicity, we'll omit those here.
+	reader := blobDownloadResponse.Body(nil)
+	downloadData, err := ioutil.ReadAll(reader)
 	handle(err)
-	err = reader.Close()
-	handle(err)
-	if data != downloadedData.String() {
-		handle(errors.New("downloaded data doesn't match uploaded data"))
+	if string(downloadData) != uploadData {
+		handle(errors.New("Uploaded data should be same as downloaded data"))
 	}
 
-	// ===== 3. list blobs =====
-	// The ListBlobs and ListContainers APIs return two channels, a values channel, and an errors channel.
-	// You should enumerate on a range over the values channel, and then check the errors channel, as only ONE value will ever be passed to the errors channel.
-	// The AutoPagerTimeout defines how long it will wait to place into the items channel before it exits & cleans itself up. A zero time will result in no timeout.
-	pager := container.ListBlobsFlat(nil)
 
-	for pager.NextPage(ctx) {
+	if err = reader.Close(); err != nil {
+		handle(err)
+		return
+	}
+
+	// ===== 3. List blobs =====
+	// List methods returns a pager object which can be used to iterate over the results of a paging operation.
+	// To iterate over a page use the NextPage(context.Context) to fetch the next page of results.
+	// PageResponse() can be used to iterate over the results of the specific page.
+	// Always check the Err() method after paging to see if an error was returned by the pager. A pager will return either an error or the page of results.
+	pager := containerClient.ListBlobsFlat(nil)
+	for pager.NextPage(context.TODO()) {
 		resp := pager.PageResponse()
-
-		for _, v := range resp.ContainerListBlobFlatSegmentResult.Segment.BlobItems {
+		for _, v := range resp.Segment.BlobItems {
 			fmt.Println(*v.Name)
 		}
 	}
@@ -196,12 +202,12 @@ Examples
 		handle(err)
 	}
 
-	// Delete the blob we created earlier.
-	_, err = blockBlob.Delete(ctx, nil)
+	// Delete the blob.
+	_, err = blockBlobClient.Delete(context.TODO(), nil)
 	handle(err)
 
-	// Delete the container we created earlier.
-	_, err = container.Delete(ctx, nil)
+	// Delete the container.
+	_, err = containerClient.Delete(context.TODO(), nil)
 	handle(err)
 */
